Rework GalaxyUtilityContainerManager to avoid a Galaxy-> Tool Shed import.
