+++ lib/tool_shed/galaxy_install/dependency_display.py
@@ -3,7 +3,7 @@ import os
 
 from galaxy import util
 from galaxy.tool_shed.util.common_util import parse_repository_dependency_tuple
-from tool_shed.galaxy_install.utility_containers import GalaxyUtilityContainerManager
+from galaxy.tool_shed.util import utility_container_manager
 
 log = logging.getLogger(__name__)
 
@@ -251,3 +251,46 @@ class DependencyDisplayer:
                         if req_name == name and req_type == type:
                             return False
         return True
+
+
+class GalaxyUtilityContainerManager(utility_container_manager.UtilityContainerManager):
+    def __init__(self, app):
+        self.app = app
+
+    def build_repository_containers(
+        self,
+        missing_repository_dependencies,
+        repository_dependencies,
+    ):
+        """
+        Return a dictionary of containers for the received repository's dependencies and readme files for
+        display during installation to Galaxy.
+        """
+        containers_dict = dict(
+            repository_dependencies=None,
+            missing_repository_dependencies=None,
+        )
+        # Some of the tool dependency folders will include links to display tool dependency information, and
+        # some of these links require the repository id.  However we need to be careful because sometimes the
+        # repository object is None.
+        try:
+            folder_id = 0
+            # Installed repository dependencies container.
+            if repository_dependencies:
+                label = "Installed repository dependencies"
+                folder_id, repository_dependencies_root_folder = self.build_repository_dependencies_folder(
+                    folder_id=folder_id, repository_dependencies=repository_dependencies, label=label, installed=True
+                )
+                containers_dict["repository_dependencies"] = repository_dependencies_root_folder
+            # Missing repository dependencies container.
+            if missing_repository_dependencies:
+                folder_id, missing_repository_dependencies_root_folder = self.build_repository_dependencies_folder(
+                    folder_id=folder_id,
+                    repository_dependencies=missing_repository_dependencies,
+                    label="Missing repository dependencies",
+                    installed=False,
+                )
+                containers_dict["missing_repository_dependencies"] = missing_repository_dependencies_root_folder
+        except Exception as e:
+            log.debug(f"Exception in build_repository_containers: {str(e)}")
+        return containers_dict
